[orabos] A/B root partitions, Third for persistence
We could either have two completely separate paritions,
and copy things manually around, or this.
Btrfs offers the option create subvolumes for partitioning
the data further up.
Otherwise we would have to statically determine the size
of the home, var, and var/tmp partitions (or at least:
home and var).
Home should be fairly limited, so that would not be an insurmountable
problem.

As the VMs will (eventually) only run with non-local storage, the
partition will only be used for our data, which should be practically
ephemeral, and is only persisted for a faster life-cycle.

Brtfs also would offer us the option to use snapshots for rollback,
partition the paths further, or and makes it easy to add more paths,
in case of need.

This includes the read-only /usr directory, now disabled.
Adding a normal a/b partition doubles the risk of having
chosen the wrong size and having to redeploy from scratch to
repartition it.